Adds view separation, in preparation for server modularization.
