+++ launch_physics/__init__.py
@@ -1,0 +1,4 @@
+from flask import Flask
+app = Flask(__name__)
+
+import launch_physics.views
