--- static/html/about.html
@@ -1,13 +1,0 @@
-<!--About Page-->
-<head>
-    <link rel="stylesheet" type="text/css" href="main.css">
-    <link href='https://fonts.googleapis.com/css?family=Exo+2' rel='stylesheet' type='text/css'>
-
-</head>
-
-<div>
-  <ul>
-      <li><a href = "LaunchLogin.html"><img src = "logo.png"></a></li>
-      <a href = "About.html"><li class = "word">About</li ></a>
-  </ul>
-</div>
