--- static/html/modules.html
@@ -1,33 +1,0 @@
-<head>
-    <link rel="stylesheet" type="text/css" href="main.css">
-    <link href='https://fonts.googleapis.com/css?family=Exo+2' rel='stylesheet' type='text/css'>
-</head>
-
-<div class = "top">
-  <ul>
-      <a href = "LaunchLogin.html"><li class = "word">Logout</li></a>
-      <a href = "Badges.html"><li class = "word">Badges</li></a>
-      <a href = "MyAccount.html"><li class = "word">My Account</li></a>
-      <a href = "LaunchCommandCentral.html"><li class = "word">Home</li></a>
-  </ul>
-</div>
-
-<div class = "dummy"></div>
-
-<div>
-    <h1>(Track Name)</h1>
-      <h4>(Module Name)</h4>
-      <p>(Sections)</p>
-      <p>(Excercises)</p>
-      <h4>(Module Name)</h4>
-      <p>(Sections)</p>
-      <p>(Excercises)</p>
-      <h4>(Module Name)</h4>
-      <p>(Sections)</p>
-      <p>(Excercises)</p>
-      <h4>(Module Name)</h4>
-      <p>(Sections)</p>
-      <p>(Excercises)</p>
-      <h4>(Module Name)</h4>
-      <p>(Sections)</p>
-      <p>(Excercises)</p>
