--- static/html/register.html
@@ -1,36 +1,0 @@
-<head>
-    <link rel="stylesheet" type="text/css" href="main.css">
-    <link href='https://fonts.googleapis.com/css?family=Exo+2' rel='stylesheet' type='text/css'>
-</head>
-
-<div>
-  <ul>
-      <li><a href = "LaunchLogin.html"><img src = "logo.png"></a></li>
-      <a href = "About.html"><li class = "word">About</li ></a>
-  </ul>
-</div>
-
-<div>
-    <div class = "login"> <!-- Box around all of the login -->
-        <h1>LAUNCH! Register</h1>
-        <p>Learn the beauty of physics!</p>
-        <form action="demo_form.asp">
-          Username: <input type="text" name = "username" maxlength = "50"><br>
-          Password: <input type= "password" name =" password" maxlength = "50"><br>
-          Check Password: <input type = "password" name = "check password" maxlength = "50"><br>
-          Email: <input type = "email" name = "email"><br>
-          Age:
-          <form>
-            <input type="radio" name="age" value="underthirteen" checked>Under 13<br>
-            <input type="radio" name="age" value="between">13-17<br>
-            <input type="radio" name="age" value="above">Above 17
-          </form>
-          <input type="submit" value="Submit" class = "button">
-        </form>
-        <a href = "LaunchLogin.html"><div>Click here to login</div></a>
-    </div>
-</div>
-
-<div>
-    <!--bottom of website crap-->
-</div>
